refactor(shape): integrate batch/download/compute extraction\n\n- use services/batch/BatchSessionManager across code/tests\n- remove legacy services/BatchSessionManager.ts\n- re-export batch manager and download factory from services index\n- scoped typecheck passes for shape-plugin
